pvme_spreadsheet/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
google-auth==2.23.3
    # via
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.1.0
    # via gspread
gspread==5.11.3
    # via -r requirements.in
idna==3.4
    # via requests
oauthlib==3.2.2
    # via requests-oauthlib
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
requests==2.31.0
    # via requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
urllib3==2.0.6
    # via requests
